split to Container/View components
