+++ src/filterConfig.js
